requirements.txt
# XVNNN-RAT v1.0.0 - Dependencies
# Author: Ali Zafar (alizafarbati@gmail.com)

# Core dependencies (required)
pynput>=1.7.0
mss>=9.0.0
Pillow>=9.0.0
psutil>=5.9.0

# Optional dependencies (recommended)
opencv-python>=4.5.0
pyaudio>=0.2.11

# Build tools
pyinstaller>=5.0.0

# Note: Install with:
# pip3 install -r requirements.txt --break-system-packages